# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chinook API",
        "contact": {
            "name": "Testsmith",
            "email": "info@testsmith.io"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local environment"
        }
    ],
    "paths": {
...
        "/genres": {
            "get": {
                "tags": [
                    "Genre"
                ],
                "summary": "Retrieve all genres",
                "operationId": "get-all-genres",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page, default=1",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All genre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GenreResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Genre"
                ],
                "summary": "New genre",
                "operationId": "store-genre",
                "requestBody": {
                    "description": "Genre object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An Genre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenreResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/genres/{genreId}": {
            "get": {
                "tags": [
                    "Genre"
                ],
                "summary": "Retrieve a specific genre",
                "operationId": "get-genre",
                "parameters": [
                    {
                        "name": "genreId",
                        "in": "path",
                        "description": "The genreId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An Genre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenreResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            },
            "put": {
                "tags": [
                    "Genre"
                ],
                "summary": "Update genre",
                "operationId": "update-genre",
                "parameters": [
                    {
                        "name": "genreId",
                        "in": "path",
                        "description": "The genreId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Genre object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An Genre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GenreResponse"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Genre"
                ],
                "summary": "Remove genre",
                "operationId": "destroy-genre",
                "parameters": [
                    {
                        "name": "genreId",
                        "in": "path",
                        "description": "The genreId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/genres/search": {
            "get": {
                "tags": [
                    "Genre"
                ],
                "summary": "Retrieve specific genres matching the search query",
                "operationId": "search-genre",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "The query parameter in path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All matching genres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GenreResponse"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GenreRequest": {
                "title": "GenreRequest",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Class Genre",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "GenreResponse": {
                "title": "GenreResponse",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}